move components into separate folders for a cleaner structure
