+++ client/src/components/account/Register/Register.vue
@@ -1,11 +1,13 @@
 <template>
     <CreateAccount v-if="stage === 'create'" v-on:success="stage = 'photos'"></CreateAccount>
-    <UploadPhotos v-else></UploadPhotos>
+    <UploadPhotos v-else-if="stage === 'photos'" v-on:success="stage = 'compose'"></UploadPhotos>
+    <ComposeProfile v-else-if="stage === 'compose'"></ComposeProfile>
 </template>
 
 <script>
 import CreateAccount from './CreateAccount'
 import UploadPhotos from './PhotoUpload'
+import ComposeProfile from './ComposeProfile'
 
 export default {
     data() {
@@ -15,7 +17,8 @@ export default {
     },
     components: {
         CreateAccount,
-        UploadPhotos
+        UploadPhotos,
+        ComposeProfile,
     }
 }
 </script>
